fix : Post.tsx을 Type 폴더로 이동, Detail 컴포넌트 세분화
