+++ 신동현/src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Home, SignUp, SignUpDetail } from "./pages";
 import GlobalStyle from "./styles/Global";
-import Detail from "./pages/Detail";
+import Detail from "./pages/Detail/Detail";
 import Write from "./pages/Write";
 import Mypage from "./pages/Mypage";
 function App() {
